# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CreditTech API documentation",
        "contact": {
            "email": "myself.ronin@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
        "/applications": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "Fetch all application from database",
                "operationId": "getApplications",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            },
            "post": {
                "tags": [
                    "Application"
                ],
                "summary": "Add a application to product",
                "operationId": "addApplication",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            }
        },
        "/applications/{id}": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "Show a application detail",
                "operationId": "showApplication",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Application not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Application"
                ],
                "summary": "Delete a application to Application table",
                "operationId": "deleteApplication",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Application with that id not found"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Fetch all product from database",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add a product",
                "operationId": "addProduct",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show a product detail",
                "operationId": "showProduct",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Product not found with that id"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product from Product table",
                "operationId": "deleteProduct",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Product not found with that id"
                    }
                }
            }
        },
        "/product/{id}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update a product details",
                "operationId": "updateProduct",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Product not found with that id"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Fetch all the user",
                "operationId": "usersAll",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add a user to Users table",
                "operationId": "addUser",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show a user",
                "operationId": "showUser",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user to Users table",
                "operationId": "updateUser",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user to Users table",
                "operationId": "deleteUser",
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            }
        }
    }
}